+++ components/ui/index.ts
@@ -1,0 +1,29 @@
+import SectionTitle from './SectionTitle';
+import FormInput from './FormInput';
+import Button from './Button';
+import ViewEducationData from './ViewEducationData';
+import SaveButton from './SaveButton';
+import EditButton from './EditButton';
+import DataLists from './DataLists';
+import HeaderItem from './HeaderItem';
+import HeaderForm from './HeaderForm';
+import EducationForm from './EducationForm';
+import EducationFormSubsection from './EducationFormSubsection';
+import EducationItem from './EducationItem';
+import ResumeSection from './ResumeSection';
+
+export {
+    SectionTitle,
+    FormInput,
+    Button,
+    SaveButton,
+    ViewEducationData,
+    EditButton,
+    ResumeSection,
+    HeaderItem,
+    EducationItem,
+    DataLists,
+    HeaderForm,
+    EducationForm,
+    EducationFormSubsection,
+};