+++ hooks/useViewData/index.ts
@@ -1,0 +1,13 @@
+import { useState, useEffect } from 'react';
+
+const useViewData = (onCallValue: boolean = false) => {
+    const [viewData, setViewData] = useState<boolean>(false);
+
+    useEffect(() => {
+        setViewData(onCallValue);
+    }, [onCallValue]);
+
+    return { viewData, setViewData };
+};
+
+export default useViewData;