+++ utils/index.ts
@@ -1,5 +1,8 @@
 import makeArrayHeaderData from './makeArrayHeaderData';
 import shortenLink from './shortenLink';
-import sortArrayEducationData from './sortEducationData';
+import sortArrayData from './sortArrayData';
 
-export { makeArrayHeaderData, shortenLink, sortArrayEducationData };
+// HOCS
+import withSubmitHandler from './hoc/withSubmitFormHandler';
+
+export { makeArrayHeaderData, shortenLink, sortArrayData, withSubmitHandler };